Add unit tests for analysis, fix some issues in analysis logic

+++ main/models.py
@@ -12,6 +12,6 @@ class QuoteAnalysis(models.Model):
     total_letters = models.IntegerField()
     total_vowels = models.IntegerField()
     total_consonants = models.IntegerField()
-    average_word_length = models.DecimalField(max_digits=9, decimal_places=2,)
+    average_word_length = models.IntegerField()
     letter_repetitions = models.TextField()
     longest_words = models.TextField()
